+++ src/snowplow_signals/models/get_attributes_response.py
@@ -1,0 +1,7 @@
+from typing import Any
+
+from pydantic import BaseModel
+
+
+class GetAttributesResponse(BaseModel):
+    data: dict[str, list[Any]]