+++ test/test_signals.py
@@ -136,3 +136,45 @@ class TestSignalsApply:
         assert view_post_mock.called
         assert view_get_mock.called
         assert apply_mock.called
+
+
+class TestSignalsGetAttributes:
+    def test_get_service_attributes(self, respx_mock, signals_client):
+        api_response = {
+            "domain_userid": ["user-123"],
+            "page_views_count": [10],
+        }
+        respx_mock.post(
+            "http://localhost:8000/api/v1/get-online-attributes",
+            json__service="my_service",
+            json__entities={"domain_userid": ["user-123"]},
+        ).mock(return_value=httpx.Response(200, json=api_response))
+
+        response = signals_client.get_service_attributes(
+            name="my_service", entity="domain_userid", identifier="user-123"
+        )
+
+        assert response["domain_userid"] == "user-123"
+        assert response["page_views_count"] == 10
+
+    def test_get_view_attributes(self, respx_mock, signals_client):
+        api_response = {
+            "domain_userid": ["user-123"],
+            "page_views_count": [10],
+        }
+        respx_mock.post(
+            "http://localhost:8000/api/v1/get-online-attributes",
+            json__entities={"domain_userid": ["user-123"]},
+            json__attributes=["my_view_v1:page_views_count"],
+        ).mock(return_value=httpx.Response(200, json=api_response))
+
+        response = signals_client.get_view_attributes(
+            name="my_view",
+            version=1,
+            attributes=["page_views_count"],
+            entity="domain_userid",
+            identifier="user-123",
+        )
+
+        assert response["domain_userid"] == "user-123"
+        assert response["page_views_count"] == 10